Docs: 코드 분리
